requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX

numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
pysdf
dearpygui
packaging
scipy
lpips
imageio
torchmetrics